一覧編集画面削除方法変更

+++ src/flask_practice/views.py
@@ -5,7 +5,14 @@ storing flask views
 
 from flask import redirect, render_template, request, url_for
 from flask_wtf import FlaskForm
-from wtforms import BooleanField, FieldList, FormField, StringField, SubmitField
+from wtforms import (
+    BooleanField,
+    FieldList,
+    FormField,
+    IntegerField,
+    StringField,
+    SubmitField,
+)
 from wtforms.validators import InputRequired
 
 from flask_practice import app, db
@@ -13,6 +20,7 @@ from flask_practice.models.ranking import Ranking
 
 STRINGFIELD_ITEM_NAME = "item_name"
 STRINGFIELD_PLACE = "place"
+INTEGERFIELD_PPRICE = "price"
 STRINGFIELD_RANKING_CATEGORY = "ranking_category"
 STRINGFIELD_ITEM_IMAGE = "item_image"
 BOOLEANFIELD_IS_DELETED = "is_deleted"
@@ -28,16 +36,16 @@ class ItemForm(FlaskForm):
 
     item_name = StringField(STRINGFIELD_ITEM_NAME, validators=[InputRequired()])
     place = StringField(STRINGFIELD_PLACE, validators=[InputRequired()])
+    price = IntegerField(INTEGERFIELD_PPRICE)
     ranking_category = StringField(STRINGFIELD_RANKING_CATEGORY)
     item_image = StringField(STRINGFIELD_ITEM_IMAGE)
-    is_deleted = BooleanField(BOOLEANFIELD_IS_DELETED, default="")
 
 
 class RankingForm(FlaskForm):
     """ranking list form for edit"""
 
     items = FieldList(FormField(ItemForm, "Item"))
-    submit = SubmitField("送信")
+    submitform = SubmitField("送信")
     addline = SubmitField("Add new line")
 
     def update_self(self):
@@ -52,13 +60,13 @@ class RankingForm(FlaskForm):
         if read_form_data["addline"]:
             updated_list.append({})
 
-        # update item list removing deleted rows
-        if read_form_data["submit"]:
-            updated_list_delete_row = updated_list
-            updated_list = []
-            for item in updated_list_delete_row:
-                if not item.get(BOOLEANFIELD_IS_DELETED):
-                    updated_list.append(item)
+        # # update item list removing deleted rows
+        # if read_form_data["submitform"]:
+        #     updated_list_delete_row = updated_list
+        #     updated_list = []
+        #     for item in updated_list_delete_row:
+        #         if not item.get(BOOLEANFIELD_IS_DELETED):
+        #             updated_list.append(item)
 
         # reload the form from the modified data
         read_form_data["items"] = updated_list
@@ -114,19 +122,26 @@ def ranking_edit():
     """Edit ranking page"""
     items = Ranking.query.all()
     ranking_form = RankingForm(items=items)
+    print("validate_on_submit before")
+    print(ranking_form.data)
 
-    if RankingForm().validate_on_submit():
-        # This reloads the form with the editted values/added rows/deleted rows
+    # This reloads the form with the editted values/added rows/deleted rows
+
+    if ranking_form.data["addline"]:
         ranking_form.update_self()
+        print(ranking_form.data)
+        return render_template("edit-ranking.html", ranking_form=ranking_form)
 
+    if ranking_form.data["submitform"] and RankingForm().validate_on_submit():
+        print("validate_on_submit after")
         # Clear DB
         db.session.query(Ranking).delete()
 
         # update DB with entered values in ranking form
-        print(ranking_form.data)
         for item in ranking_form.data["items"]:
             ranking = Ranking(
                 item_name=item.get(STRINGFIELD_ITEM_NAME),
+                price=item.get(INTEGERFIELD_PPRICE),
                 place=item.get(STRINGFIELD_PLACE),
                 ranking_category=item.get(STRINGFIELD_RANKING_CATEGORY),
                 item_image=item.get(STRINGFIELD_ITEM_IMAGE),
